refactor(libr): split sys.rs into focused submodules
sys.rs had grown to 2400 lines covering R discovery, initialization,
console I/O, the spinner, error state, askpass and interrupt handling.
Move each concern into its own submodule and re-export from sys.rs so
that every existing path keeps resolving. Pure code movement, no
behavior change.
